+++ src/azure-cli/azure/cli/command_modules/network/_template_builder.py
@@ -34,9 +34,10 @@ def _build_frontend_ip_config(cmd, name, public_ip_id=None, subnet_id=None, priv
     return frontend_ip_config
 
 
-# pylint: disable=too-many-locals
+# pylint: disable=too-many-locals, too-many-statements
 def build_application_gateway_resource(cmd, name, location, tags, sku_name, sku_tier, capacity, servers, frontend_port,
-                                       private_ip_address, private_ip_allocation, cert_data, cert_password,
+                                       private_ip_address, private_ip_allocation,
+                                       cert_data, cert_password, key_vault_secret_id,
                                        cookie_based_affinity, http_settings_protocol, http_settings_port,
                                        http_listener_protocol, routing_rule_type, public_ip_id, subnet_id,
                                        connection_draining_timeout, enable_http2, min_capacity, zones,
@@ -89,6 +90,15 @@ def build_application_gateway_resource(cmd, name, location, tags, sku_name, sku_
         }
         if cert_password:
             ssl_cert['properties']['password'] = "[parameters('certPassword')]"
+
+    if key_vault_secret_id:
+        http_listener['properties'].update({'SslCertificate': {'id': ssl_cert_id}})
+        ssl_cert = {
+            'name': ssl_cert_name,
+            'properties': {
+                'keyVaultSecretId': key_vault_secret_id,
+            }
+        }
 
     backend_http_settings = {
         'name': http_settings_name,